+++ FlorezApp/Sources/General/MainViewController.swift
@@ -1,0 +1,22 @@
+//
+//  MainViewController.swift
+//  FlorezApp
+//
+//  Created by Bodgar Espinosa Miranda on 03/03/24.
+//
+
+import UIKit
+
+class MainViewController: UIViewController {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(hideKeyboard))
+        view.addGestureRecognizer(tapGesture)
+    }
+    
+    @objc func hideKeyboard() {
+        view.endEditing(true)
+    }
+
+}
